refactor: move `modules`-dependent code out of `params` (only builds; does not work)
