+++ params/config.go
@@ -215,6 +215,7 @@ var (
 // ChainConfig is stored in the database on a per block basis. This means
 // that any network, identified by its genesis block, can have its own
 // set of configuration options.
+// TODO document the need to call InitChainConfig (DO NOT MERGE)
 type ChainConfig struct {
 	ChainID *big.Int `json:"chainId"` // chainId identifies the current chain and is used for replay protection
 
@@ -373,12 +374,6 @@ func (r *Rules) PredicatersExist() bool {
 func (r *Rules) PredicaterExists(addr common.Address) bool {
 	_, PredicaterExists := r.Predicaters[addr]
 	return PredicaterExists
-}
-
-// IsPrecompileEnabled returns whether precompile with [address] is enabled at [timestamp].
-func (c *ChainConfig) IsPrecompileEnabled(address common.Address, timestamp uint64) bool {
-	config := c.getActivePrecompileConfig(address, timestamp)
-	return config != nil && !config.IsDisabled()
 }
 
 // CheckCompatible checks whether scheduled fork transitions have been imported
@@ -568,10 +563,11 @@ func (c *ChainConfig) checkCompatible(newcfg *ChainConfig, height *big.Int, time
 		return err
 	}
 
+	// TODO: DO NOT MERGE
 	// Check that the precompiles on the new config are compatible with the existing precompile config.
-	if err := c.CheckPrecompilesCompatible(newcfg.PrecompileUpgrades, time); err != nil {
-		return err
-	}
+	// if err := c.CheckPrecompilesCompatible(newcfg.PrecompileUpgrades, time); err != nil {
+	// 	return err
+	// }
 
 	// Check that the state upgrades on the new config are compatible with the existing state upgrade config.
 	if err := c.CheckStateUpgradesCompatible(newcfg.StateUpgrades, time); err != nil {
@@ -757,6 +753,8 @@ func (c *ChainConfig) Rules(blockNum *big.Int, timestamp uint64) Rules {
 	rules := c.rules(blockNum, timestamp)
 
 	rules.AvalancheRules = c.GetAvalancheRules(timestamp)
+
+	// TODO document the need to call InitChainRules (DO NOT MERGE)
 
 	return rules
 }
